Move some generic utilities into a separate library
Prepare for extracting these into a separate repo at some time, for use
in other of my extensions.
