+++ src/lib/source.ts
@@ -21,7 +21,7 @@ import Gio from "gi://Gio";
 import Soup from "gi://Soup";
 
 import { SourceMetadata } from "./source/metadata.js";
-import { DownloadableImage } from "./util/download.js";
+import { DownloadableImage } from "./download.js";
 
 export type { SourceMetadata } from "./source/metadata.js";
