+++ src/lib/ui/indicator.ts
@@ -33,12 +33,12 @@ import {
 import { gettext as _ } from "resource:///org/gnome/shell/extensions/extension.js";
 
 import { ImageMetadata, SourceMetadata } from "../source.js";
-import { IconLoader } from "./icons.js";
+import { IconLoader } from "../common/ui/icons.js";
 import { RefreshState } from "../services/refresh.js";
-import i18n from "../util/i18n.js";
-import { Destructible } from "../util/lifecycle.js";
+import i18n from "../common/i18n.js";
+import { Destructible } from "../common/lifecycle.js";
 import SOURCES from "../sources/metadata/sources.js";
-import { ImageFile } from "../util/download.js";
+import { ImageFile } from "../download.js";
 
 class ImageInfoSection extends PopupMenuSection {
   private readonly title: PopupMenuItem;
